refactor(cli): extract search command module (search, search:stats, reindex)
